SWDEV-505971 - change setArgument arg from uint32_t to uint64_t
We are passing this arg as an address, and memcpy complains about
overreading (8 bytes instead of 4).

+++ rocclr/device/pal/palblit.hpp
@@ -478,7 +478,7 @@ class KernelBlitManager : public DmaBlitManager {
   //! Batch memory ops- Submits batch of streamWaits and streamWrite operations.
   virtual bool batchMemOps(const void* paramArray,
                            size_t paramSize,
-                           uint32_t count) const {
+                           uint64_t count) const {
     assert(!"Unimplemented");
     return false;
   }
